+++ packages/expo-audio/src/utils/options.ts
@@ -3,7 +3,7 @@ import { Platform } from 'expo-modules-core';
 import { RecordingOptions } from '../Audio.types';
 
 export function createRecordingOptions(options: RecordingOptions) {
-  let newOptions = {
+  let commonOptions = {
     extension: options.extension,
     sampleRate: options.sampleRate,
     numberOfChannels: options.numberOfChannels,
@@ -11,15 +11,15 @@ export function createRecordingOptions(options: RecordingOptions) {
   };
 
   if (Platform.OS === 'ios') {
-    newOptions = {
-      ...newOptions,
+    commonOptions = {
+      ...commonOptions,
       ...options.ios,
     };
   } else if (Platform.OS === 'android') {
-    newOptions = {
-      ...newOptions,
+    commonOptions = {
+      ...commonOptions,
       ...options.android,
     };
   }
-  return newOptions;
+  return commonOptions;
 }