+++ packages/expo-router/plugin/src/index.ts
@@ -47,7 +47,8 @@ const withRouter: ConfigPlugin<
     extra: {
       ...config.extra,
       router: {
-        origin: false,
+        // RSC enables location origin by default because it's required for requests.
+        origin: config.experiments?.reactServerComponents ? undefined : false,
         ...config.extra?.router,
         ...props,
       },